+++ lib/src/count/count.dart
@@ -4,13 +4,14 @@ import 'package:stringr/src/split/words.dart';
 /// Extension bundling functionalities related to counting string units
 extension Count on String {
   /// Counts the characters in a string
+  @Deprecated('Use .length instead')
   int count() => length;
 
   /// Counts the graphemes in a string
   int countGrapheme() => characters.length;
 
-  /// Counts the occurences of [find] in a string
-  int countOccurences(String find) {
+  /// Counts the occurrences of [find] in a string
+  int countOccurrences(String find) {
     var count = 0;
     var matchIndex = 0;
     if (isEmpty || find.isEmpty) {
@@ -25,6 +26,10 @@ extension Count on String {
     } while (matchIndex != -1);
     return count;
   }
+
+  /// Counts the occurrences of [find] in a string
+  @Deprecated('Use countOccurrences instead (note the spelling)')
+  int countOccurences(String find) => countOccurrences(find);
 
   /// Counts the characters in a string for which predicate returns `true`
   int countWhere(bool Function(String character) predicate) {