+++ lib/src/split/words.dart
@@ -6,7 +6,7 @@ extension Words on String {
   /// Split a string according to a given pattern/regex. Splits on [" "] by
   /// default. To split according to a regex, pass the regex pattern string to
   /// [pattern] parameter
-  List<String?> words({Pattern pattern = ""}) {
+  List<String> words({Pattern pattern = ""}) {
     Pattern patternRegExp;
     if (pattern == "") {
       patternRegExp = RegExp(regexpExtendedAscii).hasMatch(this)
@@ -17,16 +17,18 @@ extension Words on String {
     }
     return RegExp(patternRegExp.toString())
         .allMatches(this)
-        .map((e) => e.group(0))
+        .map((e) => e.group(0)!)
         .toList();
   }
 
   /// Splits a string into list of single characters
-  List<String?> chars() => split('');
+  @Deprecated('Use .split(\'\') instead')
+  List<String> chars() => split('');
 
   /// Returns an array of unicode code point values
-  List<int?> codePoints() => codeUnits;
+  @Deprecated('Use .codeUnits instead')
+  List<int> codePoints() => codeUnits.toList();
 
   /// Returns an array of graphemes of the string
-  List<String?> graphemes() => characters.toList();
+  List<String> graphemes() => characters.toList();
 }